+++ sdk/servicebus/service-bus/src/session/messageSession.ts
@@ -27,7 +27,7 @@ import { BatchingReceiverLite, MinimalReceiver } from "../core/batchingReceiver"
 import { onMessageSettled, DeferredPromiseAndTimer } from "../core/shared";
 import { AbortError, AbortSignalLike } from "@azure/abort-controller";
 import { ReceiverHelper } from "../core/receiverHelper";
-import { CreateSessionReceiverOptions, MessageHandlerOptionsBase } from "../models";
+import { AcceptSessionOptions, MessageHandlerOptionsBase } from "../models";
 
 /**
  * Describes the options that need to be provided while creating a message session receiver link.
@@ -49,7 +49,7 @@ export interface CreateMessageSessionReceiverLinkOptions {
  * Describes all the options that can be set while instantiating a MessageSession object.
  */
 export type MessageSessionOptions = Pick<
-  CreateSessionReceiverOptions<"receiveAndDelete">,
+  AcceptSessionOptions<"receiveAndDelete">,
   "maxAutoRenewLockDurationInMs" | "abortSignal"
 > & {
   receiveMode?: InternalReceiveMode;